Merge reactormixins-4987
Author: exarkun
Reviewer: spiv, jesstess
Fixes: #4987

Previously the `ReactorBuilder`-based tests kept a single list of reactors
to generate unit tests for.  Change this so that the current platform is
considered when defining the list, and omit reactors which certainly cannot
be tested on the platform in use.

Fix a couple small problems revealed by this, both with PortableGtkReactor
on POSIX (where it isn't really supported, but the fixes are trivial).


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@31477 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb